Fill in new tests for other (one-to-one or many-to-many) case
While writing tests for the three different functionalites, I realized
some situations were not tested for all functionality. Add tests to
the one-to-one or many-to-many test files such that we are testing
the same set of things across the three different functionalities.

For the one-to-one case, there are fewer tests because some of
tests for the one-to-many and many-to-many cases don't apply.